requirements.txt
certifi
charset-normalizer
idna
lxml
request
urllib3
